Refactor pyresample.utils in to separate modules
