+++ pyresample/test/test_plot.py
@@ -59,29 +59,27 @@ class Test(unittest.TestCase):
 
     @unittest.skipIf(Basemap is None, "basemap is not available")
     def test_area_def2basemap(self):
-        from pyresample import plot, utils
-        area_def = utils.parse_area_file(os.path.join(os.path.dirname(__file__),
-                                                      'test_files', 'areas.cfg'), 'ease_sh')[0]
+        from pyresample import plot
+        from pyresample import parse_area_file
+        area_def = parse_area_file(os.path.join(os.path.dirname(__file__), 'test_files', 'areas.cfg'), 'ease_sh')[0]
         bmap = plot.area_def2basemap(area_def)
-        self.assertTrue(bmap.rmajor == bmap.rminor and
-                        bmap.rmajor == 6371228.0,
+        self.assertTrue(bmap.rmajor == bmap.rminor and bmap.rmajor == 6371228.0,
                         'Failed to create Basemap object')
 
     def test_plate_carreeplot(self):
-        from pyresample import plot, utils, kd_tree, geometry
-        area_def = utils.parse_area_file(os.path.join(os.path.dirname(__file__),
-                                                      'test_files', 'areas.cfg'), 'pc_world')[0]
+        from pyresample import plot, kd_tree, geometry
+        from pyresample import parse_area_file
+        area_def = parse_area_file(os.path.join(os.path.dirname(__file__), 'test_files', 'areas.cfg'), 'pc_world')[0]
         swath_def = geometry.SwathDefinition(self.lons, self.lats)
         result = kd_tree.resample_nearest(swath_def, self.tb37v, area_def,
                                           radius_of_influence=20000,
                                           fill_value=None)
-        plt = plot._get_quicklook(area_def, result, num_meridians=0,
-                                  num_parallels=0)
+        plt = plot._get_quicklook(area_def, result, num_meridians=0, num_parallels=0)
 
     def test_easeplot(self):
-        from pyresample import plot, utils, kd_tree, geometry
-        area_def = utils.parse_area_file(os.path.join(os.path.dirname(__file__),
-                                                      'test_files', 'areas.cfg'), 'ease_sh')[0]
+        from pyresample import plot, kd_tree, geometry
+        from pyresample import parse_area_file
+        area_def = parse_area_file(os.path.join(os.path.dirname(__file__), 'test_files', 'areas.cfg'), 'ease_sh')[0]
         swath_def = geometry.SwathDefinition(self.lons, self.lats)
         result = kd_tree.resample_nearest(swath_def, self.tb37v, area_def,
                                           radius_of_influence=20000,
@@ -89,9 +87,9 @@ class Test(unittest.TestCase):
         plt = plot._get_quicklook(area_def, result)
 
     def test_orthoplot(self):
-        from pyresample import plot, utils, kd_tree, geometry
-        area_def = utils.parse_area_file(os.path.join(os.path.dirname(__file__),
-                                                      'test_files', 'areas.cfg'), 'ortho')[0]
+        from pyresample import plot, kd_tree, geometry
+        from pyresample import parse_area_file
+        area_def = parse_area_file(os.path.join(os.path.dirname(__file__), 'test_files', 'areas.cfg'), 'ortho')[0]
         swath_def = geometry.SwathDefinition(self.lons, self.lats)
         result = kd_tree.resample_nearest(swath_def, self.tb37v, area_def,
                                           radius_of_influence=20000,
@@ -100,8 +98,7 @@ class Test(unittest.TestCase):
 
 
 def suite():
-    """The test suite.
-    """
+    """The test suite."""
     loader = unittest.TestLoader()
     mysuite = unittest.TestSuite()
     mysuite.addTest(loader.loadTestsFromTestCase(Test))
